+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This file is used to list changes made in each version of the AWS ParallelCluste
   - line-4.5.13 (from line-4.5.2)
   - openssh-2.11.3 (from openssh-2.10.3)
   - pyenv-4.2.3 (from pyenv-3.5.1)
+  - selinux-6.1.12 (from selinux-6.0.5)
 
 **BUG FIXES**
 
@@ -50,6 +51,7 @@ This file is used to list changes made in each version of the AWS ParallelCluste
     - openssh-2.11.3 (from openssh-2.10.3)
     - pyenv-4.2.3 (from pyenv-3.5.1)
     - selinux-6.1.12 (from selinux-6.0.5)
+    - yum-7.4.13 (from yum-7.4.0)
 
 **BUG FIXES**
 - Fix cluster creation failure with Ubuntu Deep Learning AMI on GPU instances and DCV enabled.