+++ CHANGELOG.md
@@ -14,6 +14,14 @@ This file is used to list changes made in each version of the AWS ParallelCluste
 **BUG FIXES**
 - Fix inconsistent scaling configuration after cluster update rollback when modifying the list of instance types declared in the Compute Resources.
 
+3.7.1
+------
+
+**CHANGES**
+- Upgrade Slurm to 23.02.5 (from 23.02.4).
+  - Upgrade Pmix to 4.2.6 (from 3.2.3).
+  - Upgrade libjwt to 1.15.3 (from 1.12.0).
+
 3.7.0
 ------